# src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FountTechs API Documentation",
    "version": "1.0.0",
    "description": "This is an API for storing and retrieving information about local startups in Kinshasa. These businesses' data are accessible through the API which can be used by any frontend. The API is built using Node.js and Express, and the documentation is generated using Swagger. The API provides endpoints for creating, reading, updating, and deleting startup information, as well as searching for startups based on various criteria. The documentation includes details about each endpoint, including the request parameters, response format, and example requests and responses. ",
    "author": "Baron T.",
    "last update": "2026-05-22"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter JWT token as: Bearer <token>"
    }
  },
  "paths": {
    "/startups/": {
      "get": {
        "tags": [
          "Startups Endpoints"
        ],
        "summary": "Get all startups",
        "description": "Pull all startups from the database. ",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Startups Endpoints"
        ],
        "summary": "Create a new startup",
        "description": "Insert new startup in the database. ",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "New startup data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "KongoTech"
                },
                "description": {
                  "type": "string",
                  "example": "FinTech startup providing digital payment solutions"
                },
                "industry": {
                  "type": "string",
                  "example": "FinTech"
                },
                "founders": {
                  "type": "array",
                  "example": [
                    "Baron Mobs"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "foundedYear": {
                  "type": "number",
                  "example": 2022
                },
                "location": {
                  "type": "object",
                  "properties": {
                    "ward": {
                      "type": "string",
                      "example": "Ngiri-Ngiri"
                    },
                    "stake": {
                      "type": "string",
                      "example": "Kinshasa"
                    },
                    "commune": {
                      "type": "string",
                      "example": "Ngiri-Ngiri"
                    },
                    "city": {
                      "type": "string",
                      "example": "Kinshasa"
                    },
                    "province": {
                      "type": "string",
                      "example": "Kinshasa"
                    },
                    "country": {
                      "type": "string",
                      "example": "DR Congo"
                    }
                  }
                },
                "phone": {
                  "type": "string",
                  "example": "+243810000000"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/users/": {
      "post": {
        "tags": [
          "Users Endpoints"
        ],
        "summary": "Create a new user",
        "description": "Insert a new user into the database.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "New user data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "example": "Baron"
                },
                "lastName": {
                  "type": "string",
                  "example": "Mobs"
                },
                "username": {
                  "type": "string",
                  "example": "barontech"
                },
                "email": {
                  "type": "string",
                  "example": "baron@founttech.com"
                },
                "password": {
                  "type": "string",
                  "example": "123456"
                },
                "role": {
                  "type": "string",
                  "example": "Founder"
                },
                "phone": {
                  "type": "string",
                  "example": "+243810000000"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "tags": [
          "Users Endpoints"
        ],
        "summary": "Get all users",
        "description": "Retrieve all users from the database.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  }
}